Build protoc code for viewing
By default, `tonic-build` uses a macro to generate necessary code using
specified protobuffers. I want to output all the code to a permanent
directory instead so I can see the code, but also to validate that
protobuffer code can be generated as a seperate stage so that tonic is
usable when providing generated code as a crate.