Set service status on empty service name
When health checking a gRPC server, clients can specify the empty string
in order to check the general health of the service. This is optional
for servers, so Tonic does not handle this case.

Custom logic would better serve developers, but in this case we don't
care. I'd rather not break any clients that check the empty service by
default though, so we'll intervene here.